pyproject.toml
[project]
name = "timbremetrics"
description = "Metrics to test the alignment between audio models and human timbre perception"
dynamic = ["version"]
dependencies = [
    "soundfile",
    "numpy",
    "torch",
    "torchaudio",
    "torchmetrics",
]

[project.optional-dependencies]
dev = [
    "black",
]
scripts = [
    "scipy",
    "matplotlib",
    "pyyaml",
]
examples = [
    "fadtk==1.0.1",
    "cdpam",
    "descript-audio-codec==1.0.0",
    "kymatio @ git+https://github.com/kymatio/kymatio.git@2e190d92a19ce274988590c834e978063cb0c490",
]

[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["timbremetrics"]
